% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\alias{aagi_report}
\alias{html_presentation}
\alias{short_report}
\alias{knitr_report}
\title{Create a new AAGI Report using a template}
\usage{
new_report(filename, type)

aagi_report(filename)

html_presentation(filename)

short_report(filename)

knitr_report(filename)
}
\arguments{
\item{filename}{The filename of the report.}

\item{type}{The type of the report. Valid types are `aagi_report`,
`html_presentation`, `short_report`, and `knitr_report`.
Partial matching is performed.}
}
\value{
Returns the path to the created file invisibly.
}
\description{
Create a new AAGI Report using a template
}
\examples{
\dontrun{
new_report(filename = "test_report", type = "aagi_report")
}
}
