% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\alias{bh_report}
\alias{sagi_report}
\alias{html_presentation}
\title{Create a new Biometry Hub Report using a template}
\usage{
new_report(filename, type)

bh_report(filename)

sagi_report(filename)

html_presentation(filename)
}
\arguments{
\item{filename}{The filename of the report}

\item{type}{The type of the report. Valid types are `biometry_hub_report`, `sagi_report`, `html_presentation`.}
}
\value{
Returns the path to the created file invisibly.
}
\description{
Create a new Biometry Hub Report using a template
}
\examples{
\dontrun{
new_report(filename = "test_report", type = "biometry_hub_report")
}
}
